Move YAML config normalization into schema layer
