+++ src/config.rs
@@ -5,22 +5,18 @@ use std::fs;
 use std::path::{Path, PathBuf};
 
 use ignore::gitignore::{Gitignore, GitignoreBuilder};
-use regex::Regex;
-use saphyr::{LoadableYamlNode, MappingOwned, ScalarOwned, YamlOwned};
-use serde::de::DeserializeOwned;
+use saphyr::{LoadableYamlNode, ScalarOwned, YamlOwned};
 use toml::Value as TomlValue;
 
 use crate::config_schema::{
     FixRuleName as TomlFixRuleName, FixableRuleSelector as TomlFixableRuleSelector,
-    ForbidSetting, IndentSequencesSetting, NewLinesType, NormalizedConfig,
-    NormalizedFixConfig, QuoteType, QuotedStringsRequiredModeForValidation,
-    SpacesSetting, TomlConfig, TruthyAllowedValue, load_ignore_from_files,
-    load_ignore_patterns, normalize_toml_config, parse_string_items,
-    parse_toml_config_str, quoted_strings_required_mode_from_yaml_value,
-    validate_key_ordering_patterns, validate_quoted_strings_semantics,
-    validate_toml_config, yaml_owned_to_toml_value, yaml_value_matches_toml_type,
+    NormalizedConfig, NormalizedFixConfig, TomlConfig, load_ignore_from_files,
+    load_ignore_patterns, normalize_toml_config, normalize_yaml_config,
+    parse_toml_config_str, validate_toml_config, yaml_owned_to_toml_value,
 };
 use crate::{conf, decoder};
+
+pub use crate::config_schema::RuleLevel;
 
 /// Abstraction over environment/filesystem to enable full test coverage.
 /// Minimal environment abstraction used by tests to cover file system and env-var behavior.
@@ -133,29 +129,11 @@ pub struct YamlLintConfig {
 
 const DEFAULT_YAML_FILE_PATTERNS: [&str; 3] = ["*.yaml", "*.yml", ".yamllint"];
 
-const TRUTHY_ALLOWED_VALUES_DISPLAY: &str = "['YES', 'Yes', 'yes', 'NO', 'No', 'no', 'TRUE', 'True', 'true', 'FALSE', 'False', 'false', 'ON', 'On', 'on', 'OFF', 'Off', 'off']";
-
 #[derive(Debug, Clone, Default)]
 struct RuleFilter {
     patterns: Vec<String>,
     from_files: Vec<String>,
     matcher: Option<Gitignore>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum RuleLevel {
-    Error,
-    Warning,
-}
-
-impl RuleLevel {
-    fn parse(value: &str) -> Option<Self> {
-        match value {
-            "error" => Some(Self::Error),
-            "warning" => Some(Self::Warning),
-            _ => None,
-        }
-    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -556,7 +534,7 @@ impl YamlLintConfig {
         if let Some(extends) = doc.as_mapping_get("extends") {
             cfg.apply_extends(extends, envx, base_dir)?;
         }
-        cfg.apply_normalized_config(normalize_yaml_doc(doc)?);
+        cfg.apply_normalized_config(normalize_yaml_config(doc)?);
 
         Ok(cfg)
     }
@@ -850,78 +828,6 @@ fn build_rule_filter(
     Ok(())
 }
 
-fn normalize_yaml_doc(doc: &YamlOwned) -> Result<NormalizedConfig, String> {
-    let ignore = doc.as_mapping_get("ignore");
-    let ignore_from_file = doc.as_mapping_get("ignore-from-file");
-    if ignore.is_some() && ignore_from_file.is_some() {
-        return Err(
-            "invalid config: ignore and ignore-from-file keys cannot be used together"
-                .to_string(),
-        );
-    }
-
-    let yaml_file_patterns = match doc.as_mapping_get("yaml-files") {
-        Some(yaml_files) => {
-            let Some(seq) = yaml_files.as_sequence() else {
-                return Err(
-                    "invalid config: yaml-files should be a list of file patterns"
-                        .to_string(),
-                );
-            };
-            let mut patterns = Vec::with_capacity(seq.len());
-            for item in seq {
-                let Some(pattern) = item.as_str() else {
-                    return Err(
-                        "invalid config: yaml-files should be a list of file patterns"
-                            .to_string(),
-                    );
-                };
-                patterns.push(pattern.to_owned());
-            }
-            Some(patterns)
-        }
-        None => None,
-    };
-
-    let locale = match doc.as_mapping_get("locale") {
-        Some(locale) => {
-            let Some(locale) = locale.as_str() else {
-                return Err("invalid config: locale should be a string".to_string());
-            };
-            Some(locale.to_owned())
-        }
-        None => None,
-    };
-
-    if doc.as_mapping_get("fix").is_some() {
-        return Err(
-            "invalid config: fix is only supported in TOML configuration".to_string(),
-        );
-    }
-
-    let mut normalized = NormalizedConfig {
-        ignore_patterns: ignore.map(load_ignore_patterns).transpose()?,
-        ignore_from_files: ignore_from_file.map(load_ignore_from_files).transpose()?,
-        yaml_file_patterns,
-        locale,
-        ..NormalizedConfig::default()
-    };
-
-    if let Some(rules) = doc.as_mapping_get("rules")
-        && let Some(map) = rules.as_mapping()
-    {
-        for (key, value) in map {
-            let Some(name) = key.as_str() else {
-                continue;
-            };
-            validate_rule_value(name, value)?;
-            normalized.rules.insert(name.to_owned(), value.clone());
-        }
-    }
-
-    Ok(normalized)
-}
-
 fn determine_rule_level(node: &YamlOwned) -> Option<RuleLevel> {
     if let Some(s) = node.as_str() {
         return if s == "disable" {
@@ -938,480 +844,17 @@ fn determine_rule_level(node: &YamlOwned) -> Option<RuleLevel> {
     node.as_mapping()
         .and_then(|map| {
             map.iter().find_map(|(key, value)| {
-                (key.as_str() == Some("level"))
-                    .then(|| value.as_str().and_then(RuleLevel::parse))
+                if key.as_str() != Some("level") {
+                    return None;
+                }
+                Some(if value.as_str() == Some("warning") {
+                    RuleLevel::Warning
+                } else {
+                    RuleLevel::Error
+                })
             })
         })
-        .flatten()
         .or(Some(RuleLevel::Error))
-}
-
-fn validate_rule_value(name: &str, value: &YamlOwned) -> Result<(), String> {
-    if let Some(text) = value.as_str() {
-        return match text {
-            "enable" | "disable" => Ok(()),
-            _ => Err(format!(
-                "invalid config: rule '{name}' should be 'enable', 'disable', or a mapping"
-            )),
-        };
-    }
-
-    if value.as_bool().is_some() {
-        return Ok(());
-    }
-
-    if let Some(map) = value.as_mapping() {
-        if name == "quoted-strings" {
-            validate_quoted_strings_rule(map)?;
-            return Ok(());
-        }
-
-        for (key, val) in map {
-            if handle_common_rule_key(name, key, val)? {
-                continue;
-            }
-            validate_specific_rule_option(name, key, val)?;
-        }
-        return Ok(());
-    }
-
-    Err(format!(
-        "invalid config: rule '{name}' should be 'enable', 'disable', or a mapping"
-    ))
-}
-
-fn handle_common_rule_key(
-    rule: &str,
-    key: &YamlOwned,
-    val: &YamlOwned,
-) -> Result<bool, String> {
-    if key.as_str() == Some("level") {
-        let Some(level_text) = val.as_str() else {
-            return Err(format!(
-                "invalid config: rule '{rule}' level should be \"error\" or \"warning\""
-            ));
-        };
-        if RuleLevel::parse(level_text).is_none() {
-            return Err(format!(
-                "invalid config: rule '{rule}' level should be \"error\" or \"warning\""
-            ));
-        }
-        return Ok(true);
-    }
-
-    if key.as_str() == Some("ignore") {
-        load_ignore_patterns(val)?;
-        return Ok(true);
-    }
-
-    if key.as_str() == Some("ignore-from-file") {
-        load_ignore_from_files(val)?;
-        return Ok(true);
-    }
-
-    Ok(false)
-}
-
-fn validate_specific_rule_option(
-    rule: &str,
-    key: &YamlOwned,
-    val: &YamlOwned,
-) -> Result<(), String> {
-    match rule {
-        "anchors" => validate_scalar_rule_option(
-            "anchors",
-            key,
-            val,
-            &[
-                "forbid-undeclared-aliases",
-                "forbid-duplicated-anchors",
-                "forbid-unused-anchors",
-            ],
-            &[],
-        ),
-        "braces" => validate_brace_like_option("braces", key, val),
-        "brackets" => validate_brace_like_option("brackets", key, val),
-        "document-end" => {
-            validate_scalar_rule_option("document-end", key, val, &["present"], &[])
-        }
-        "document-start" => {
-            validate_scalar_rule_option("document-start", key, val, &["present"], &[])
-        }
-        "empty-lines" => validate_scalar_rule_option(
-            "empty-lines",
-            key,
-            val,
-            &[],
-            &["max", "max-start", "max-end"],
-        ),
-        "commas" => validate_scalar_rule_option(
-            "commas",
-            key,
-            val,
-            &[],
-            &["max-spaces-before", "min-spaces-after", "max-spaces-after"],
-        ),
-        "comments" => validate_scalar_rule_option_allow_non_string(
-            "comments",
-            key,
-            val,
-            &["require-starting-space", "ignore-shebangs"],
-            &["min-spaces-from-content"],
-        ),
-        "new-lines" => validate_new_lines_option(key, val),
-        "octal-values" => validate_scalar_rule_option(
-            "octal-values",
-            key,
-            val,
-            &["forbid-implicit-octal", "forbid-explicit-octal"],
-            &[],
-        ),
-        "float-values" => validate_scalar_rule_option(
-            "float-values",
-            key,
-            val,
-            &[
-                "require-numeral-before-decimal",
-                "forbid-scientific-notation",
-                "forbid-nan",
-                "forbid-inf",
-            ],
-            &[],
-        ),
-        "empty-values" => validate_scalar_rule_option(
-            "empty-values",
-            key,
-            val,
-            &[
-                "forbid-in-block-mappings",
-                "forbid-in-flow-mappings",
-                "forbid-in-block-sequences",
-            ],
-            &[],
-        ),
-        "key-duplicates" => validate_scalar_rule_option(
-            "key-duplicates",
-            key,
-            val,
-            &["forbid-duplicated-merge-keys"],
-            &[],
-        ),
-        "hyphens" => {
-            validate_scalar_rule_option("hyphens", key, val, &[], &["max-spaces-after"])
-        }
-        "truthy" => validate_truthy_option(key, val),
-        "key-ordering" => validate_key_ordering_option(key, val),
-        "indentation" => validate_indentation_option(key, val),
-        "line-length" => validate_scalar_rule_option(
-            "line-length",
-            key,
-            val,
-            &[
-                "allow-non-breakable-words",
-                "allow-non-breakable-inline-mappings",
-            ],
-            &["max"],
-        ),
-        "trailing-spaces" => Err(unknown_rule_option("trailing-spaces", key)),
-        "comments-indentation" => Err(unknown_rule_option("comments-indentation", key)),
-        _ => Ok(()),
-    }
-}
-
-fn validate_brace_like_option(
-    rule: &str,
-    key: &YamlOwned,
-    val: &YamlOwned,
-) -> Result<(), String> {
-    let Some(name) = key.as_str() else {
-        return Err(unknown_rule_option(rule, key));
-    };
-
-    match name {
-        "forbid" => {
-            if yaml_value_matches_toml_type::<ForbidSetting>(val) {
-                Ok(())
-            } else {
-                Err(format!(
-                    "invalid config: option \"forbid\" of \"{rule}\" should be bool or \"non-empty\""
-                ))
-            }
-        }
-        "min-spaces-inside" => validate_int_option(val, rule, "min-spaces-inside"),
-        "max-spaces-inside" => validate_int_option(val, rule, "max-spaces-inside"),
-        "min-spaces-inside-empty" => {
-            validate_int_option(val, rule, "min-spaces-inside-empty")
-        }
-        "max-spaces-inside-empty" => {
-            validate_int_option(val, rule, "max-spaces-inside-empty")
-        }
-        _ => Err(unknown_rule_option(rule, key)),
-    }
-}
-
-fn validate_scalar_rule_option(
-    rule: &str,
-    key: &YamlOwned,
-    val: &YamlOwned,
-    bool_options: &[&str],
-    int_options: &[&str],
-) -> Result<(), String> {
-    let Some(name) = key.as_str() else {
-        return Err(unknown_rule_option(rule, key));
-    };
-
-    if bool_options.contains(&name) {
-        return validate_bool_option(val, rule, name);
-    }
-    if int_options.contains(&name) {
-        return validate_int_option(val, rule, name);
-    }
-
-    Err(unknown_rule_option(rule, key))
-}
-
-fn validate_scalar_rule_option_allow_non_string(
-    rule: &str,
-    key: &YamlOwned,
-    val: &YamlOwned,
-    bool_options: &[&str],
-    int_options: &[&str],
-) -> Result<(), String> {
-    if key.as_str().is_none() {
-        return Ok(());
-    }
-
-    validate_scalar_rule_option(rule, key, val, bool_options, int_options)
-}
-
-fn validate_new_lines_option(key: &YamlOwned, val: &YamlOwned) -> Result<(), String> {
-    if key.as_str() != Some("type") {
-        return Err(unknown_rule_option("new-lines", key));
-    }
-
-    validate_typed_option_value::<NewLinesType>(
-        val,
-        "invalid config: option \"type\" of \"new-lines\" should be in ('unix', 'dos', 'platform')",
-    )
-}
-
-fn validate_truthy_option(key: &YamlOwned, val: &YamlOwned) -> Result<(), String> {
-    match key.as_str() {
-        Some("allowed-values") => {
-            validate_typed_option_value::<Vec<TruthyAllowedValue>>(
-                val,
-                format!(
-                    "invalid config: option \"allowed-values\" of \"truthy\" should only contain values in {TRUTHY_ALLOWED_VALUES_DISPLAY}"
-                ),
-            )
-        }
-        Some("check-keys") => validate_bool_option(val, "truthy", "check-keys"),
-        _ => Err(unknown_rule_option("truthy", key)),
-    }
-}
-
-fn validate_key_ordering_option(
-    key: &YamlOwned,
-    val: &YamlOwned,
-) -> Result<(), String> {
-    match key.as_str() {
-        Some("ignored-keys") => {
-            let patterns = parse_string_items(
-                val,
-                "invalid config: option \"ignored-keys\" of \"key-ordering\" should contain regex strings",
-                |text| vec![text.to_owned()],
-            )?;
-            validate_key_ordering_patterns(&patterns)
-        }
-        _ => Err(unknown_rule_option("key-ordering", key)),
-    }
-}
-
-fn validate_indentation_option(key: &YamlOwned, val: &YamlOwned) -> Result<(), String> {
-    match key.as_str() {
-        Some("spaces") => validate_typed_option_value::<SpacesSetting>(
-            val,
-            "invalid config: option \"spaces\" of \"indentation\" should be in (<class 'int'>, 'consistent')",
-        ),
-        Some("indent-sequences") => {
-            validate_typed_option_value::<IndentSequencesSetting>(
-                val,
-                "invalid config: option \"indent-sequences\" of \"indentation\" should be in (<class 'bool'>, 'whatever', 'consistent')",
-            )
-        }
-        Some("check-multi-line-strings") => {
-            validate_bool_option(val, "indentation", "check-multi-line-strings")
-        }
-        _ => Err(unknown_rule_option("indentation", key)),
-    }
-}
-
-fn validate_quoted_strings_rule(map: &MappingOwned) -> Result<(), String> {
-    let mut state = QuotedStringsValidationState::default();
-    for (key, val) in map {
-        if handle_common_rule_key("quoted-strings", key, val)? {
-            continue;
-        }
-        validate_quoted_strings_option(key, val, &mut state)?;
-    }
-    state.finish()
-}
-
-#[derive(Default)]
-struct QuotedStringsValidationState {
-    required: Option<QuotedStringsRequiredModeForValidation>,
-    extra_required_count: Option<usize>,
-    extra_allowed_count: Option<usize>,
-}
-
-impl QuotedStringsValidationState {
-    fn finish(&self) -> Result<(), String> {
-        let extra_required = self
-            .extra_required_count
-            .map(|count| vec![String::new(); count]);
-        let extra_allowed = self
-            .extra_allowed_count
-            .map(|count| vec![String::new(); count]);
-        validate_quoted_strings_semantics(
-            self.required
-                .unwrap_or(QuotedStringsRequiredModeForValidation::True),
-            extra_required.as_deref(),
-            extra_allowed.as_deref(),
-        )
-    }
-}
-
-fn validate_quoted_strings_option(
-    key: &YamlOwned,
-    val: &YamlOwned,
-    state: &mut QuotedStringsValidationState,
-) -> Result<(), String> {
-    match key.as_str() {
-        Some("quote-type") => validate_quote_type_option(val),
-        Some("required") => validate_required_option(val, state),
-        Some("extra-required") => validate_regex_list_option(
-            val,
-            "extra-required",
-            &mut state.extra_required_count,
-        ),
-        Some("extra-allowed") => validate_regex_list_option(
-            val,
-            "extra-allowed",
-            &mut state.extra_allowed_count,
-        ),
-        Some("allow-quoted-quotes") => {
-            validate_bool_option(val, "quoted-strings", "allow-quoted-quotes")
-        }
-        Some("check-keys") => validate_bool_option(val, "quoted-strings", "check-keys"),
-        _ => Err(unknown_rule_option("quoted-strings", key)),
-    }
-}
-
-fn validate_quote_type_option(val: &YamlOwned) -> Result<(), String> {
-    validate_typed_option_value::<QuoteType>(
-        val,
-        "invalid config: option \"quote-type\" of \"quoted-strings\" should be in ('any', 'single', 'double')",
-    )
-}
-
-fn validate_required_option(
-    val: &YamlOwned,
-    state: &mut QuotedStringsValidationState,
-) -> Result<(), String> {
-    if let Some(required) = quoted_strings_required_mode_from_yaml_value(val) {
-        state.required = Some(required);
-        Ok(())
-    } else {
-        Err(
-            "invalid config: option \"required\" of \"quoted-strings\" should be in (True, False, 'only-when-needed')"
-                .to_string(),
-        )
-    }
-}
-
-fn validate_regex_list_option(
-    val: &YamlOwned,
-    option_name: &str,
-    count_slot: &mut Option<usize>,
-) -> Result<(), String> {
-    let Some(seq) = val.as_sequence() else {
-        return Err(quoted_strings_regex_type_error(option_name));
-    };
-    *count_slot = Some(seq.len());
-    validate_regex_strings(
-        val,
-        &quoted_strings_regex_type_error(option_name),
-        |text, err| {
-            format!(
-                "invalid config: regex \"{text}\" in option \"{option_name}\" of \"quoted-strings\" is invalid: {err}"
-            )
-        },
-    )
-}
-
-fn validate_bool_option(
-    val: &YamlOwned,
-    rule_name: &str,
-    option_name: &str,
-) -> Result<(), String> {
-    validate_option_type::<bool>(val, rule_name, option_name, "bool")
-}
-
-fn validate_int_option(
-    val: &YamlOwned,
-    rule_name: &str,
-    option_name: &str,
-) -> Result<(), String> {
-    validate_option_type::<i64>(val, rule_name, option_name, "int")
-}
-
-fn validate_option_type<T: DeserializeOwned>(
-    val: &YamlOwned,
-    rule_name: &str,
-    option_name: &str,
-    expected_type: &str,
-) -> Result<(), String> {
-    if yaml_value_matches_toml_type::<T>(val) {
-        Ok(())
-    } else {
-        Err(format!(
-            "invalid config: option \"{option_name}\" of \"{rule_name}\" should be {expected_type}"
-        ))
-    }
-}
-
-fn validate_typed_option_value<T: DeserializeOwned>(
-    val: &YamlOwned,
-    error_message: impl Into<String>,
-) -> Result<(), String> {
-    if yaml_value_matches_toml_type::<T>(val) {
-        Ok(())
-    } else {
-        Err(error_message.into())
-    }
-}
-
-fn unknown_rule_option(rule: &str, key: &YamlOwned) -> String {
-    let key_name = describe_rule_option_key(key);
-    format!("invalid config: unknown option \"{key_name}\" for rule \"{rule}\"")
-}
-
-fn validate_regex_strings(
-    val: &YamlOwned,
-    type_error: &str,
-    invalid_regex: impl Fn(&str, regex::Error) -> String,
-) -> Result<(), String> {
-    let values = parse_string_items(val, type_error, |text| vec![text.to_owned()])?;
-    for text in values {
-        Regex::new(&text).map_err(|err| invalid_regex(&text, err))?;
-    }
-    Ok(())
-}
-
-fn quoted_strings_regex_type_error(option_name: &str) -> String {
-    format!(
-        "invalid config: option \"{option_name}\" of \"quoted-strings\" should only contain values in [<class 'str'>]"
-    )
 }
 
 fn resolve_extend_path(
@@ -1458,23 +901,6 @@ fn deep_merge_yaml_owned(dst: &mut YamlOwned, src: &YamlOwned) {
         }
     } else {
         *dst = src.clone();
-    }
-}
-
-fn describe_rule_option_key(key: &YamlOwned) -> String {
-    match (
-        key.as_integer(),
-        key.as_floating_point(),
-        key.as_bool(),
-        key.is_null(),
-        key.as_str(),
-    ) {
-        (Some(num), _, _, _, _) => num.to_string(),
-        (None, Some(float), _, _, _) => float.to_string(),
-        (None, None, Some(flag), _, _) => flag.to_string(),
-        (None, None, None, true, _) => "None".to_string(),
-        (None, None, None, false, Some(text)) => text.to_owned(),
-        _ => format!("{key:?}"),
     }
 }
